moved tool utility functions definitions to common utils.py file to remove code redundancy
